Refactor files to standalone file
